requirements.txt
Django==1.11.13
django-authtools==1.6.0
django-model-utils==3.1.2
djangorestframework==3.8.2
pytz==2018.4
